Move import to fix wasm build
